vj_server/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
python-dotenv
